change struct bme680_dev to ease platform support
Remove dev_id member of struct bme680_dev  and introduce a context
pointer member. The context pointer member provides greater flexibility
for clients to store information required to implement the
read/write/delay functions. For example, from Linux userspace, it's
necessary to store the file descriptor of the open I2C bus and the I2C
address of the bme680.